requirements.txt
flask
openai
python-dotenv
gunicorn
langchain 
faiss-cpu
tiktoken
babel
pymongo